+++ hooks/authUser.js
@@ -13,7 +13,7 @@ export const RequireAuth = () => {
 
     useEffect(() => {
         if (!user) {
-            router.push('/auth/login')
+            router.push('/auth')
         }
     }, [user, router])
 }
