+++ theme/styles.js
@@ -1,7 +1,9 @@
 import { mode } from '@chakra-ui/theme-tools'
 
 export const globalStyles = {
-    initialColorMode: 'dark',
+    config: {
+        initialColorMode: 'dark',
+    },
     colors: {
         gray: {
             700: '#1f2733',
